ran notebook

+++ src/readdata.py
@@ -79,11 +79,13 @@ def to_undirected(edge_iterable, num_edges, shuffle=True):
 
 def get_clean_data(DATA_DIRECTORY, EDGE_LIST, NUM_EDGES, shuffle=True):
     if os.path.exists(DATA_DIRECTORY + EDGE_LIST + '_edges.txt'):
-        edges = np.loadtxt(DATA_DIRECTORY + EDGE_LIST + '_edges.txt')
+        edges = np.loadtxt(DATA_DIRECTORY + EDGE_LIST + '_edges.txt').astype(np.int32)
         num_nodes = len(np.unique(edges))
         neighborsMap = [[] for node in range(num_nodes)]
         node_indices = [[] for node in range(num_nodes)]
         for i in range(edges.shape[0]):
+#             neighborsMap[int(edges[i,0])].append(int(edges[i,1]))
+#             node_indices[int(edges[i,0])].append(i)
             neighborsMap[edges[i,0]].append(edges[i,1])
             node_indices[edges[i,0]].append(i)
     else:
